requirements.txt
streamlit
PyPDF2
pptx
google-generativeai
pdfplumber
tiktoken
